requirements.txt
# Install this first
pyomo[optional]

# Logging and monitoring
matplotlib 
tqdm
snakeviz
wandb
tensorboardX

# Config system
hydra-core

# ML & OR libraries
numpy
scipy
pyomo[optional]
gurobipy